backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
numpy>=1.24.0
# CLIP — install separately (not on PyPI):
# pip install git+https://github.com/openai/CLIP.git
